backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.15.0
python-multipart==0.0.6
ebaysdk==2.2.0
python-dotenv>=0.19.0
beautifulsoup4==4.12.2
httpx==0.25.1
google-cloud-vision==3.4.4
google-cloud-storage==2.13.0
pydantic>=2.4.2
python-jose==3.3.0
passlib==1.7.4
google-generativeai==0.3.1
